+++ test/pageBridgeStaleGuard.test.js
@@ -17,6 +17,10 @@ const pageBridgeSource = fs.readFileSync(
   path.join(__dirname, '../extension/src/pageBridge.js'),
   'utf8'
 );
+const binaryAssetUploaderSource = fs.readFileSync(
+  path.join(__dirname, '../extension/src/page/binaryAssetUploader.js'),
+  'utf8'
+);
 const saveStateSource = fs.readFileSync(
   path.join(__dirname, '../extension/src/page/saveState.js'),
   'utf8'
@@ -2995,6 +2999,7 @@ function createPageBridgeHarness({
   vm.runInContext(overleafProjectSnapshotSource, context, { filename: 'overleafProjectSnapshot.js' });
   vm.runInContext(treeOperationsSource, context, { filename: 'treeOperations.js' });
   vm.runInContext(snapshotRouterSource, context, { filename: 'snapshotRouter.js' });
+  vm.runInContext(binaryAssetUploaderSource, context, { filename: 'binaryAssetUploader.js' });
   vm.runInContext(writeGuardSource, context, { filename: 'writeGuard.js' });
   vm.runInContext(trackedChangesLifecycleSource, context, { filename: 'trackedChangesLifecycle.js' });
   vm.runInContext(writebackRouterSource, context, { filename: 'writebackRouter.js' });